Minor changes
 - Change the type of the variables used to hold the indexes and sizes
   of arrays from 'int' to 'size_t'

 - Change the type of the variable holding the result of the arithmetic
   operation on pointers from 'int' to 'ptrdiff_t'

 - Improve readability

 - Cosmetic

+++ src/histogram.h
@@ -1,18 +1,20 @@
 #ifndef HISTOGRAM_H_
 #define HISTOGRAM_H_
 
+#include <stddef.h>
+
 /*
  * Generates horizontal histogram whose bar heights correspond to the values
  * in the 'values' array. Bars are made of the 'c' character. Returns
  * dynamically allocated string that should be freed by the caller.
  */
-char* generate_horizontal_histogram(int values[], int values_len, char c);
+char* generate_horizontal_histogram(int values[], size_t values_len, char c);
 
 /*
  * Generates vertical histogram whose bar heights correspond to the values
  * in the 'values' array. Bars are made of the 'c' character. Returns
  * dynamically allocated string that should be freed by the caller.
  */
-char* generate_vertical_histogram(int values[], int values_len, char c);
+char* generate_vertical_histogram(int values[], size_t values_len, char c);
 
 #endif /* HISTOGRAM_H_ */
